refactor: move package from io.gdcc.spi.export to io.gdcc.export
This is to make things aligned with the groupId in use.
