+++ src/main/resources/io/gdcc/export/ddipdf/i18n.inc.xsl
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<xsl:stylesheet version="2.0" xmlns:xsl="http://www.w3.org/1999/XSL/Transform" xmlns:outline="http://worldbank.org/toolkit/cdrom/outline" exclude-result-prefixes="outline">
+	<xsl:param name="language-code" select="'en'"/>
+	<xsl:variable name="msg" select="document(concat('file:/io/gdcc/spi/export/ddipdf/','messages_',$language-code,'.properties.xml'))"/>
+</xsl:stylesheet>
