--- src/main/resources/io/gdcc/spi/export/ddipdf/messages_ja.properties.xml
@@ -1,161 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<properties>
-<comment>Generated by Properties2Xml on Tue Feb 13 13:55:43 EST 2007</comment>
-<entry key="Valid">有効な</entry>
-<entry key="Frequency_table_not_shown">度数表(Frequency table)は表示されません</entry>
-<entry key="Derivation">由来</entry>
-<entry key="discrete">不連続な</entry>
-<entry key="Data_Collection_Mode">データ収集モード</entry>
-<entry key="Other_Acknowledgments">その他の確認事項</entry>
-<entry key="Identification">識別番号</entry>
-<entry key="NesstarExplorer_Click_here_to_access_nesstar_file">データファイルにアクセスするにはここをクリックしてください</entry>
-<entry key="Untitled">無題</entry>
-<entry key="Value">値</entry>
-<entry key="Percentage">割合</entry>
-<entry key="Weighting">ウェイティング</entry>
-<entry key="Primary_Investigators">第一次調査官</entry>
-<entry key="Toolkit_Blurb_PreText">この文書はToolkitを使用して作られました</entry>
-<entry key="Sampling">サンプリング</entry>
-<entry key="Cases">ケース</entry>
-<entry key="Access_Conditions">アクセス条件</entry>
-<entry key="Source">情報源</entry>
-<entry key="Modalities">様相</entry>
-<entry key="Rights_and_Disclaimer">権利及び声明文</entry>
-<entry key="Definition">定義</entry>
-<entry key="Estimates_of_Sampling_Error">サンプルエラーの見積もり</entry>
-<entry key="Data_Files">データファイル</entry>
-<entry key="Geographic_Coverage">地理的な適用範囲</entry>
-<entry key="April">４月</entry>
-<entry key="Mean">平均</entry>
-<entry key="Metadata_Production">メタデータ製作</entry>
-<entry key="Data_Collection">データ収集</entry>
-<entry key="Missing_Data">損失データ</entry>
-<entry key="Scripts_and_programs">スクリプトおよびプログラム</entry>
-<entry key="Variables">可変的</entry>
-<entry key="Interviewer_instructions">面接者の指示</entry>
-<entry key="Funding_Agencies">出資機関</entry>
-<entry key="November">１１月</entry>
-<entry key="Version">バージョン</entry>
-<entry key="Universe">共通の</entry>
-<entry key="Contributors">貢献者</entry>
-<entry key="Access_Authority">アクセス権限</entry>
-<entry key="Data_Processing_and_Appraisal">データ処理、評価</entry>
-<entry key="Scope">範囲, 領域</entry>
-<entry key="Administrative_documents">管理用文章</entry>
-<entry key="Contacts">連絡先</entry>
-<entry key="Label">ラベル</entry>
-<entry key="Technical_documents">技術的な文書</entry>
-<entry key="Decimals">小数点</entry>
-<entry key="Type">タイプ</entry>
-<entry key="Literal_question">文字の質問</entry>
-<entry key="Concepts">概念</entry>
-<entry key="Range">範囲</entry>
-<entry key="Abstract">要約</entry>
-<entry key="June">６月</entry>
-<entry key="Supervision">監督</entry>
-<entry key="Other_Forms_of_Data_Appraisal">その他ファーマットのデータ評価</entry>
-<entry key="References">参照</entry>
-<entry key="Accessibility">アクセス、入手法</entry>
-<entry key="Data_Collection_Dates">データ収集日</entry>
-<entry key="Data_Editing">データ編集</entry>
-<entry key="Questionnaires">質問</entry>
-<entry key="Reports_and_analytical_documents">レポートおよび分析的な文書</entry>
-<entry key="Valid_cases">有効な場合</entry>
-<entry key="Copyright">コピーライト</entry>
-<entry key="Documentation">書類</entry>
-<entry key="Deviations_from_Sample_Design">サンプルデザインによる偏差</entry>
-<entry key="Publishers">発行者</entry>
-<entry key="February">２月</entry>
-<entry key="Dataset_contains">データセットに含まれる</entry>
-<entry key="Acknowledgments">確認事項</entry>
-<entry key="Continuous">連続的な</entry>
-<entry key="Standard_deviation">標準偏差</entry>
-<entry key="Variables_Description">変数の記述</entry>
-<entry key="Producer">製作者</entry>
-<entry key="Production_Date">製作日</entry>
-<entry key="NesstarExplorer_Overview">” Explorer”によってデータファイルを参照することも一般的に使えわれている統計データフォーマットに変換。抽出することも可能です</entry>
-<entry key="Discrete">不連続性</entry>
-<entry key="Group">グループ</entry>
-<entry key="July">７月</entry>
-<entry key="Filename">ファイルの名前</entry>
-<entry key="Cases_Abbreviation">ケース</entry>
-<entry key="Name">名前</entry>
-<entry key="Statistical_tables">統計表</entry>
-<entry key="December">１２月</entry>
-<entry key="Subjects">主題, 内容</entry>
-<entry key="Processing_Checks">工程監査</entry>
-<entry key="NesstarExplorer_Blurb_PostText">ソフト</entry>
-<entry key="Interviewers_instructions">面接者への指示</entry>
-<entry key="Table_of_Contents">目録</entry>
-<entry key="Document_Information">書類の情報</entry>
-<entry key="Subgroups">サブグループ</entry>
-<entry key="Keywords">キーワード
-</entry>
-<entry key="groups">グループ</entry>
-<entry key="Weight">ウェイト</entry>
-<entry key="Files_Description">ファイルの詳細</entry>
-<entry key="Notes">メモ</entry>
-<entry key="Data_Collection_Notes">データ収集メモ</entry>
-<entry key="files">ファイル</entry>
-<entry key="continuous">継続的な</entry>
-<entry key="Disclaimer">声明文</entry>
-<entry key="Content">内容</entry>
-<entry key="variables">変数</entry>
-<entry key="Other_Producers">その他の製作者</entry>
-<entry key="Producers_and_Sponsors">製作者とスポンサー</entry>
-<entry key="Data_Cleaning_Notes">データクリーニングメモ</entry>
-<entry key="Distributors">分配者</entry>
-<entry key="Overview">概略</entry>
-<entry key="Citation_Requirements">引用する場合の必要条件</entry>
-<entry key="September">９月</entry>
-<entry key="Category">カテゴリー</entry>
-<entry key="Confidentiality">機密性、コンフィデンシャリティー</entry>
-<entry key="May">５月</entry>
-<entry key="Undetermined">未定</entry>
-<entry key="Structure">構造</entry>
-<entry key="file">ファイル</entry>
-<entry key="Pre-question">調査前の質問</entry>
-<entry key="Response_Rate">回答比率</entry>
-<entry key="Width">幅</entry>
-<entry key="Recoding_and_Derivation">記録と誘導</entry>
-<entry key="Series">シリー</entry>
-<entry key="October">１０月</entry>
-<entry key="Unit_of_Analysis">分析の単位</entry>
-<entry key="Data_Processing_Notes">データ処理メモ</entry>
-<entry key="Kind_of_Data">データの種類
-</entry>
-<entry key="File">ファイル</entry>
-<entry key="Time_Periods">期間</entry>
-<entry key="File_Content">ファイルの内容</entry>
-<entry key="Invalid">無効</entry>
-<entry key="Keys">キー</entry>
-<entry key="Source_of_information">情報源</entry>
-<entry key="Imputation">非難</entry>
-<entry key="Security">セキュリティー</entry>
-<entry key="NesstarExplorer_Blurb_PreText">このファイルを開けるには、無料で配布されているNesstar Explorer が必要です。</entry>
-<entry key="Other_resources">その他の資料</entry>
-<entry key="Data_Dictionnary">データ辞典</entry>
-<entry key="Information">情報</entry>
-<entry key="January">１月</entry>
-<entry key="Other_documents">その他の書類</entry>
-<entry key="Minimum">最小値</entry>
-<entry key="Scope_and_Coverage">規模及び適用範囲</entry>
-<entry key="Metadata_Producers">メタデータ製作者</entry>
-<entry key="Show_more_info">さらにインフォメーションを表示</entry>
-<entry key="Data_Collectors">データ収集者</entry>
-<entry key="Post-question">調査後の質問</entry>
-<entry key="Sampling_Procedure">サンプリングの手順</entry>
-<entry key="File_Structure">ファイルの構造</entry>
-<entry key="Variables_List">変数のリスト</entry>
-<entry key="Format">フォーマット</entry>
-<entry key="Sampling_Notes">サンプリングメモ</entry>
-<entry key="Variables_Groups">変数のグループ</entry>
-<entry key="Description">詳細</entry>
-<entry key="Categories">カテゴリー</entry>
-<entry key="Maximum">最大値</entry>
-<entry key="Depositors">デポジター、提供者、供託者</entry>
-<entry key="August">８月</entry>
-<entry key="Cover_Page">表紙</entry>
-<entry key="Weighted">ウェイトされた</entry>
-<entry key="March">３月</entry>
-</properties>
